Refactor on MetaStruct metaclass.
New we are using only the __dict__ attribute, instead of the __ordered__,
considering the PEP 520 (Preserving Class Attribute Definition Order)
introduced on Python 3.6.

More than that, the new MetaStruct implementation now inherits the 'kytos'
class attributes from base classes as if they were defined on the current class
being created, and also update the version of the attributes to the same
version of the current class.

For example, `pyof.v0x04.symmetric.hello.Hello` class direct inherits from
`pyof.v0x01.symmetric.hello.Hello`. Because of that, the normal python
inheritance would make the v0x04 version of Hello have a v0x01 Header on the
header attribute. But now, our metaclass do this check while creating the new
class and, if the version of the attribute differs from the version of the
class, then the metaclass will look for the correct version and instantiate it
as class attribute.