Created two new files the action and the flow_match the action file uses a flow_match class.
Fixing some spaces and indentations to the action file.

Delete a file to submit only one file to Kytos.

Change the name of the class to eliminate the cross-reference error from the scrutinizer. Added OPF at the beginning of the  class name.

Change the name of the class to eliminate the cross-reference error from the scrutinizer. Added OPF at the beginning of the  class name.

Change the name of the class to eliminate the cross-reference error from the scrutinizer. Added OPF at the beginning of the  class name.

Change the name of the class to eliminate the cross-reference error from the scrutinizer. Added OPF at the beginning of the  class name.

Create a new line for the parameter list that violate the length of the line.

Fix the import sorted issue.

Added Lucas Felix as a contributor

Deleted extra space before the __init__ method

Fix some indentation and spaces in comments

All indentation and spaces in Header files has been fix in relation to PEP8

Eliminated the warnings in the hello file

Changing the comment in the HelloElemVersionBitmap

Changing the indentation in the HelloElemVersionBitmap

Changing the indentation and creating spaces to the HelloElemVersionBitmap method

Rewrite the method HelloElemVersionBitmap

add some space the method HelloElemVersionBitmap: 3

add some space the method HelloElemVersionBitmap: 8 and the method __init__

Add some space to the comment in the method HelloElemVersionBitmap

Add some space to the comment in the method HelloElemVersionBitmap

Add some space to the comment in the method HelloElemVersionBitmap

Change the parameter representation to the method __init__ in the class HelloElemVersionBitmap.

Changed variable name in the HelloElemHeader

Changed the type its variable name in the HelloElemHeader

Change the name of HelloElemHeader to eliminate the cross-reference

Change the name in the all list

Fix minor changes to comply with the docstring

Fix indentation

Add a blank line to the definition in HelloElemVersionBitmap

Add a blank line to the definition in HelloElemVersionBitmap

Add a blank line to the definition in HelloElemVersionBitmap to fix the indentation

Add a newlines, spaces to fix the indentation

Add a newlines, spaces to fix the indentation

Add a newlines, spaces to fix the indentation

Add a newlines, spaces to fix the indentation

Fixing some indentations

Fixed the cross-reference issue.

Fixed the cross-reference issue by adding (:class:`~pyof.v0x05.hello.*`).

Changed OpenFlow port to 6653 (#905)

Openflow version was changed from 6633 to 6653.

updating versions of Jinja2 and urlib3

Fixed the cross-reference issue by adding another (:class:`~pyof.v0x05.hello.HelloElemHeader`)

Fixed the cross-reference issue by adding to v0x04 (:class:`~pyof.v0x04.hello.HelloElemHeader`)

Fixed line spaces and endings.

Added () between the class specification.

Fixed some comments in the description.

Fixing the cross-reference issue returning to the original name.

Fixing and indentation.